// src/types/pegawai.ts
export interface Pegawai {
    id: string;
    nama_pegawai: string;
    jenis_pegawai: string;
    nip: string;
    pangkat_golongan: string;
    jabatan_dinas: string;
    jabatan_keuangan: string;
    bidang: string;
    jenis_kelamin: string;
    is_pejabat?: boolean;
    created_at: string;
    updated_at: string;
}